requirements.txt
wikipedia
google-cloud-texttospeech
